requirements.txt
faiss-cpu
numpy
scikit-learn
sentence-transformers
fastapi
uvicorn
python-multipart
pydantic
tqdm
beautifulsoup4
streamlit
requests
pandas
loguru